+++ PetaPoco.Tests.Integration/Databases/MSSQLMsData/MssqlMsDataQueryLinqTests.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using PetaPoco.Tests.Integration.Models.MSSQL;
 using Shouldly;
 using Xunit;
 
@@ -7,6 +8,8 @@ namespace PetaPoco.Tests.Integration.Databases.MSSQLMsData
     [Collection("MssqlMsData")]
     public class MssqlMsDataQueryLinqTests : BaseQueryLinqTests
     {
+        #region Test Data
+
         private readonly StorePerson _storePerson = new StorePerson
         {
             Id = Guid.NewGuid(),
@@ -14,12 +17,16 @@ namespace PetaPoco.Tests.Integration.Databases.MSSQLMsData
             Name = "Peta"
         };
 
+        #endregion
+
         public MssqlMsDataQueryLinqTests()
             : base(new MssqlMsDataDBTestProvider())
         {
         }
 
         [Fact]
+        [Trait("Issue", "#242")]
+        [Trait("DBFeature", "Schema")]
         public void Exists_GivenPrimaryKeyMatchingOneRecordAndPocoWithSchema_ShouldBeTrue()
         {
             var pk = DB.Insert(_storePerson);
@@ -27,32 +34,24 @@ namespace PetaPoco.Tests.Integration.Databases.MSSQLMsData
         }
 
         [Fact]
+        [Trait("Issue", "#242")]
+        [Trait("DBFeature", "Schema")]
         public void Exists_GivenSqlStringMatchingOneRecordAndPocoWithSchema_ShouldBeTrue()
         {
             DB.Insert(_storePerson);
             DB.Exists<StorePerson>($"WHERE {DB.Provider.EscapeSqlIdentifier("Age")} = @0", 18).ShouldBeTrue();
         }
 
-        [Fact(DisplayName = "Support the older syntax of starting with a WHERE clause.")]
+        [Fact(DisplayName = "Exists: Support the older syntax of starting with a WHERE clause.")]
         [Trait("Category", "Regression")]
+        [Trait("Issue", "#237")]
+        [Trait("Issue", "#238")]
+        [Trait("Issue", "#242")]
+        [Trait("DBFeature", "Schema")]
         public void Exists_Regression_GivenSqlStringMatchingOneRecordAndPocoWithSchema_ShouldBeTrue()
         {
             DB.Insert(_storePerson);
             DB.Exists<StorePerson>($"{DB.Provider.EscapeSqlIdentifier("Age")} = @0", 18).ShouldBeTrue();
-        }
-
-        [TableName("store.People")]
-        [PrimaryKey("Id", AutoIncrement = false)]
-        public class StorePerson
-        {
-            [Column]
-            public Guid Id { get; set; }
-
-            [Column(Name = "FullName")]
-            public string Name { get; set; }
-
-            [Column]
-            public long Age { get; set; }
         }
     }
 }
