+++ PetaPoco.Tests.Integration/Models/MSSQL/StorePerson.cs
@@ -1,0 +1,18 @@
+﻿using System;
+
+namespace PetaPoco.Tests.Integration.Models.MSSQL
+{
+    [TableName("store.People")]
+    [PrimaryKey("Id", AutoIncrement = false)]
+    public class StorePerson
+    {
+        [Column]
+        public Guid Id { get; set; }
+
+        [Column(Name = "FullName")]
+        public string Name { get; set; }
+
+        [Column]
+        public long Age { get; set; }
+    }
+}
